requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
dnspython==2.4.1
email-validator==2.0.0.post2
executing==1.2.0
Flask==2.3.2
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
idna==3.4
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.6
WTForms==3.0.1
